type cleanup

+++ crates/brontes-classifier/src/test_utils.rs
@@ -150,7 +150,7 @@ async fn get_traces_with_meta(
     (traces, header, metadata)
 }
 
-fn get_coinbase_transfer(builder: Address, action: &Action) -> Option<u128> {
+fn get_coinbase_transfer(builder: Address, action: &Action) -> Option<u64> {
     match action {
         Action::Call(action) => {
             if action.to == builder {
